pyproject.toml
[project]
name = "graph-extractor-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.24.0",
    "opencv-python-headless>=4.8.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "plotly>=5.18.0",
    "pytesseract>=0.3.10",
    "scikit-image>=0.21.0",
    "streamlit>=1.28.0",
    "xlsxwriter>=3.1.0",
]
